requirements.txt
fastapi
google-api-python-client
google-auth
google-auth-oauthlib
pydantic
rich
uvicorn
